refactor: 금액을 BigDecimal로 전환하고 단가 스냅샷을 총액 스냅샷으로 교체
Phase 1 (3/7). Double → BigDecimal(precision=19, scale=2) 전환.
전환 과정에서 2단계의 orderedPrice(단가) 설계를 orderedAmount(총액)로 되돌린다.

왜 단가가 아니라 총액인가
- scale 2에서 40,000을 3개에 결제하면 단가 13,333.33이 되고,
  저장하는 순간 "40,000"이라는 사실이 소실된다 (13,333.33 x 3 = 39,999.99)
- "마지막 취소에서 잔여 보정" 대안은 성립하지 않는다. 보정하려면 원래 총액을 알아야 하는데
  그 정보는 취소 시점이 아니라 저장 시점에 이미 사라져 있다.
  보정하려면 총액을 저장해야 하고, 그것이 곧 총액 스냅샷이다
- scale 4 상향도 해법이 아니다. 7로 나누는 조합이면 같은 잔여가 생긴다 —
  유리수 나눗셈의 잔여는 scale로 없앨 수 없고 줄어들 뿐이다
- 단가 스냅샷이 실무 표준인 이유는 주문 건마다 행이 분리돼 단가 x 수량이 결제액과 일치하기 때문이다.
  SPEC 3절이 재주문 시 수량 누적을 계약으로 요구해 여러 주문이 한 행에 병합되는 이 구조에서는
  총액이 올바른 표현이며, 표준 이탈이 아니라 스펙이 강제한 구조에 맞춘 적응이다

부수 효과: 2단계의 가중평균 계산이 통째로 사라졌다 (orderedAmount += 결제액)

환불 계산
- 전량 취소: orderedAmount 전액. 나눗셈이 없어 반올림이 개입하지 않는다
- 부분 취소: 곱한 뒤 나눔(반올림 1회) + RoundingMode.DOWN
- 남은 총액은 재계산이 아니라 차감 — 잔여가 잔액에 남아 다음 취소로 이월된다

DOWN을 고른 근거 (경계 케이스 실측)
- HALF_UP도 합계는 정확하고 잔여가 음수가 되지도 않는다
  (orderedAmount가 scale 2 값이라 그 이하를 scale 2로 반올림해도 초과 불가)
- 다만 HALF_UP은 개별 환불이 정확한 몫을 초과할 수 있다
  (7개/10,000의 6번째 취소가 1,428.58, 정확한 몫은 1,428.57…)
- DOWN은 각 취소가 몫을 넘지 않고 잔여를 마지막에 몰아준다

검증
- 85,000을 7개에 결제(85,000/7이 나눠떨어지지 않음) 후 1개씩 7회 취소 → 원금 정확히 복귀
- 가격 변경 시나리오: 15,000x2 주문 → 50,000 인상 → 1개 취소 → 985,000
- E2E 6단계 + 회귀 19개 항목 전부 통과

API 영향 — JSON 숫자 표기가 1000000.0 → 1000000.00으로 바뀐다.
필드명이 아니라 값 표기라 계약 위반은 아니다. DECISIONS.md 3절의
"직렬화 결과 동일"이라는 당초 예상은 실측으로 정정했다.

문서: DECISIONS.md 2·3절 갱신, 8절의 리팩터링 전 클래스명·로그인 응답 서술 정정,
      JOURNAL.md에 "문제 지점을 한 단계 잘못 짚으면 해법이 통째로 틀린다" 기록

+++ src/main/java/com/sk/skala/shopapi/data/dto/OrderItemDto.java
@@ -1,5 +1,7 @@
 package com.sk.skala.shopapi.data.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,6 +18,6 @@ public class OrderItemDto {
 
 	private Long productId;
 	private String productName;
-	private Double productPrice;
+	private BigDecimal productPrice;
 	private Integer quantity;
 }
